create frame extraction module
