+++ scripts/sequitur.js
@@ -1,6 +1,6 @@
 const fs = require("mz/fs"),
-      path = require('path'),
       { execSync } = require("child_process"),
+      ef = require("./extractFrames"),
       cf = require('./compareFrames'),
       sa = require('./sampleAudio');
 
@@ -43,46 +43,13 @@ if (!fs.existsSync(dir))
 (async() => {
     // load data
     const wave = await sa.wave(audioFile, framerate),
-          numOfFrames = await loadFrames(),
+          numOfFrames = await ef.frames(videoFile, resolution),
           diffs = await cf.diffs(numOfFrames);
 
     // sequence and encode
     const seq = sequence(numOfFrames, wave, framerate, maxLevel, minOffset);
     encode(seq, resolution, framerate, audioFile, preview);
 })()
-
-
-async function loadFrames() {
-    let numOfFrames = countFrames();
-
-    if (!numOfFrames) {
-        extractFrames(videoFile, resolution);
-        numOfFrames = countFrames();
-    }
-
-    return numOfFrames;
-}
-
-function countFrames() {
-    const dir = 'temp/frames/';
-    if (fs.existsSync(dir)) {
-        files = fs.readdirSync(dir);
-        if (files.length < 2) return false;
-        else {
-            const frames = files.filter(el => path.extname(el) === '.jpg');
-            return frames.length;
-        }
-    }
-    else return false;
-}
-
-function extractFrames(file, res) {
-    const dir = 'temp/frames/';
-    if (!fs.existsSync(dir))
-        fs.mkdirSync(dir)
-
-    execSync("ffmpeg -i " + file.replace(' ','\\ ') + " -vf scale=-1:" + res + " -qscale:v 2 temp/frames/%d.jpg -y");
-}
 
 function sequence(numOfFrames, wave, fps, max, min) {
     let seq = [],
